[Loom] Propagate dependent forwarding tuples atomically
Type propagation previously refined RegionBranch yields and CFG block
arguments independently. That could specialize sibling layout values
while leaving dependent view types behind, producing invalid successor
signatures and discarding otherwise usable common refinements.

Drive structured and CFG forwarding from the fact owner’s collective
target type, project it across every source tuple, and validate
dependent SSA references under a single transaction before committing.
Reuse retained CFG predecessor edges and index definition-backed value
remaps so wide tuples stay linear without rescanning IR.

+++ loom/src/loom/transforms/cfg/cfg_simplify.c
@@ -1038,7 +1038,7 @@ static bool loom_cfg_simplify_types_equal_after_map(
   loom_type_value_remap_t remap = {
       .source_values = map->source_values,
       .target_values = map->target_values,
-      .count = map->count,
+      .count = (uint16_t)map->count,
   };
   return loom_type_equal_after_value_remap(module, source_type, target_type,
                                            &remap);
